Moved contents of container_types to builtins and replaced `make_tuple` and friends with `autograd.builtins.tuple` etc
