+++ autograd/convenience_wrappers.py
@@ -8,7 +8,6 @@ import warnings
 from .wrap_util import unary_to_nary
 from .tracer import primitive
 from .core import make_vjp as _make_vjp, make_jvp as _make_jvp, defvjp_argnum, vspace
-from .container_types import make_tuple
 
 import autograd.numpy as np
