+++ tests/test_tuple.py
@@ -3,7 +3,7 @@ from __future__ import print_function
 import autograd.numpy as np
 import autograd.numpy.random as npr
 from autograd.test_util import check_grads
-from autograd.container_types import make_tuple
+from autograd import tuple as ag_tuple
 from autograd import grad
 npr.seed(1)
 
@@ -49,7 +49,7 @@ def test_nested_higher_order():
     def outer_fun(x):
         def inner_fun(y):
             return y[0] * y[1]
-        return np.sum(np.sin(np.array(grad(inner_fun)(make_tuple(x,x)))))
+        return np.sum(np.sin(np.array(grad(inner_fun)(ag_tuple((x,x))))))
 
     check_grads(outer_fun)(5.)
     check_grads(grad(outer_fun))(10.)
